+++ app/Livewire/Sponsoring/QuickBackerAdd.php
@@ -2,11 +2,17 @@
 
 namespace App\Livewire\Sponsoring;
 
+use App\Facade\NotificationMessage;
+use App\Models\Member;
 use App\Models\Sponsoring\Backer;
+use App\Models\Sponsoring\Contract;
+use App\Models\Sponsoring\Period;
 use Livewire\Component;
 
 class QuickBackerAdd extends Component
 {
+    public ?Period $period;
+
     public string $name;
 
     public string $country;
@@ -15,8 +21,11 @@ class QuickBackerAdd extends Component
 
     public string $zip;
 
-    public function mount(): void
+    public string $member_id;
+
+    public function mount(?Period $period): void
     {
+        $this->period = $period;
         $this->country = config('app.country_code_default');
     }
 
@@ -35,8 +44,27 @@ class QuickBackerAdd extends Component
             'zip' => ['nullable', 'integer'],
         ]);
 
-        Backer::query()->create($validated);
+        $isContractCreated = false;
+        /** @var Backer $backer */
+        /** @var Member $member */
+        $backer = Backer::query()->create($validated);
+        if($this->period && $this->member_id && $member = Member::find($this->member_id)) {
+
+            $contract = new Contract();
+            $contract->backer()->associate($backer);
+            $contract->period()->associate($this->period);
+            $contract->member()->associate($member);
+            $contract->save();
+            $isContractCreated = true;
+        }
         $this->customReset();
+
+        $message = __('A new backer was successfully created.');
+        if($isContractCreated) {
+            $message .= ' '. __('And a contract with the given member was added.');
+        }
+        NotificationMessage::addSuccessNotificationMessage($message);
+
         $this->dispatch('member-contract-has-changed');
     }