+++ resources/views/sponsoring/period-member-assignment.blade.php
@@ -13,7 +13,7 @@
             JSON.parse('{{\App\Facade\NotificationMessage::popNotificationMessagesJson()}}'))"
          x-data="{showOnlyMemberWithAssignment:false}">
         <div class="bg-white overflow-hidden shadow-sm sm:rounded-lg mb-3">
-            <livewire:sponsoring.quick-backer-add/>
+            <livewire:sponsoring.quick-backer-add :period="$period"/>
         </div>
 
         <div class="flex flex-wrap justify-center items-center gap-2 mb-3">
@@ -31,7 +31,6 @@
                 @forelse(\App\Models\Member::getAllFiltered()->get() as $member)
                     @php
                         /** @var \App\Models\Member $member */
-                        /** @var \App\Models\Sponsoring\Contract $contract */
                     @endphp
                     @if($previousPeriod)
                         <livewire:sponsoring.member-backer-assignment wire:key="{{$member->id}}"