+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 ### Added
 - collapse/fold all feature for period member assignment
 - show only assigned checkbox for period member assignment
+- a member can be selected now on the quick backer add component, where a new contract will be created.
 ### Changed
 - public files url are now the file controller too
 - show loading icon on admin files view